cleanup: split prepared query bindings
